gen-ai-course/14_ai_projects/market_intelligence_agent/pyproject.toml
[project]
name = "market-intelligence-agent"
version = "0.1.0"
description = "Multi-agent market intelligence platform using LangChain and Azure AI"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "azure-search-documents>=11.4.0",
    "azure-storage-blob>=12.19.0",
    "azure-identity>=1.15.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
